# nest-backend/src/users/dto/update-user-profile-request.dto.ts
import { ApiProperty } from "@nestjs/swagger";

export default class UpdateUserProfileRequestDto {
  @ApiProperty({ required: false })
  name?: string;

  @ApiProperty({ required: false })
  bio?: string;
}